changed todo routes
