+++ routes.js
@@ -1,7 +1,7 @@
 const router = require('express').Router()
 const { requireAuth } = require('./middleware/auth')
 const { home, login, logout } = require('./controllers/authController')
-const { protected, unprotected } = require('./main_routes/allRoutes')
+const { protected, unprotected } = require('./main_routes/middlewareRoutes')
 
 router.get('/', home)
 router.get('/login', login)
@@ -11,4 +11,4 @@ router.get('/logout', logout)
 router.get('/protected', requireAuth, protected)
 router.get('/unprotected', unprotected)
 
-module.exports = router
+module.exports = router;
